Move ZoneConfig and ZoneRuntime to zone module
Relocates zone-specific dataclasses from controller.py to zone.py
where they belong with other zone-related definitions.
